FEAT: Cập nhật entity bảng giỏ hàng, vật tư, vật tư mới:  - SoLuong Int => Decimal(18,5)  - DonGia Int => Decimal(18.5)

+++ ESPlatform.QRCode.IMS.Core/DTOs/GioHang/Requests/CreatedCartSupplyRequest.cs
@@ -2,7 +2,7 @@
 
 public class CreatedCartSupplyRequest
 {
-    public int SoLuong { get; set; } = 0;
+    public decimal SoLuong { get; set; } = 0;
     public string ThongSoKyThuat { get; set; } = string.Empty;
     public string GhiChu { get; set; } = string.Empty;
     public bool IsSystemSupply { get; set; } = true;
