+++ ESPlatform.QRCode.IMS.Core/DTOs/KiemKe/Responses/SupplyListResponseItem.cs
@@ -12,7 +12,7 @@ public class SupplyListResponseItem
     
     public string Image { get; set; } = string.Empty;
 
-    public int DonGia { get; set; } = 0; 
+    public decimal DonGia { get; set; } = 0; 
     
     public string ThongSoKyThuat { get; set; } = string.Empty;
