+++ ESPlatform.QRCode.IMS.Domain/Entities/QlvtMuaSamVatTuNew.cs
@@ -18,7 +18,7 @@ public partial class QlvtMuaSamVatTuNew
 
     public string? XuatXu { get; set; }
 
-    public int? DonGia { get; set; }
+    public decimal? DonGia { get; set; }
     
-    public int? SoLuong { get; set; }
+    public decimal? SoLuong { get; set; }
 }
